Albumlist & audiolist now displaying items on first boot/fixed xml bug blocking gridview text
In album_recyclerview_album_layout.xml, the title of an album wouldn't properly display if the albumart didn't cover the entire cell. This was fixed by creating a new relativelayout below the original to hold the text.

Albumlist & Audiolist both boot and fill with items on boot, however they do not update live. Need to reload the class for them to fill with items.